+++ lib/src/directives/maxlength_directive.dart
@@ -1,0 +1,17 @@
+import 'dart:html';
+import 'package:angular/angular.dart';
+
+@Directive(selector: '[customMaxlength]')
+class MaxlengthDirective {
+  @Input()
+  int customMaxlength;
+  MaxlengthDirective(Element el) {
+    InputElement inputEl = el;
+    inputEl.onKeyPress.listen((e) {
+      if (e.keyCode != 8 && inputEl.value.length == customMaxlength) {
+        e.preventDefault();
+        return false;
+      }
+    });
+  }
+}